+++ packages/grid/src/features/standard-table/hooks/UseTableResetWhenNewData.ts
@@ -3,23 +3,17 @@ import {
   useStandardTableActions,
   useStandardTableId
 } from "./UseStandardTableConfig";
-import { getReducerIdFor } from "../redux/ReducerIdFactory";
 
 export const useTableResetWhenNewData = <TItem>(
   items: Array<TItem> | undefined
 ) => {
   const tableId = useStandardTableId();
   const {
-    actions: {
-      selectedIds: { clearSelectedIds }
-    },
+    actions: { clearSelection },
     dispatch
   } = useStandardTableActions();
 
   useEffect(() => {
-    dispatch({
-      reducerId: getReducerIdFor(tableId, "selectedIds"),
-      action: clearSelectedIds()
-    });
-  }, [items, dispatch, clearSelectedIds, tableId]);
+    dispatch(clearSelection());
+  }, [items, dispatch, clearSelection, tableId]);
 };