+++ packages/redux/src/features/higher-order-reducers/record-object-reducer/record-object-reducer.ts
@@ -1,4 +1,4 @@
-import { Reducer } from "react";
+import { Reducer } from "redux";
 import {
   RecordObjectAction,
   RecordObjectClearRecordAction,